+++ CHANGELOG.md
@@ -10,11 +10,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - code coverage
+- additional tests
 
 ### Fixed
 
 - clippy lints
 - readme link to documentation
+
+### Removed
+
+- Worker::key method
 
 ## [0.0.2] - 2023-11-22